requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
langsmith==0.2.3
marshmallow==3.23.2
momento==1.25.0
momento-wire-types==0.119.5
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.57.4
orjson==3.10.12
packaging==24.2
propcache==0.2.1
protobuf==4.25.5
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
six==1.17.0
slack_bolt==1.21.3
slack_sdk==3.33.5
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
